requirements.txt
aiohttp>=3.9.0
dnspython>=2.4.0
rich>=13.7.0
click>=8.1.0
python-whois>=0.8.0
cryptography>=41.0.0
requests>=2.31.0
